Fixed Phpstan level 3 errors and ran cs-fixer

+++ src/Feature/Logger/Builder/Monolog/GelfBuilder.php
@@ -13,7 +13,7 @@ final class GelfBuilder implements MonologBuilderInterface
     private ?string $vhost;
     private ?string $host;
     private ?int $port;
-    private ?int $timeout;
+    private ?float $timeout;
     private ?string $queue;
     private ?string $channel;
     private iterable $formatters;
